requirements.txt
Flask==3.1.3
slack-bolt==1.20.0
psycopg2-binary==2.9.11
python-dotenv==1.0.1
langchain==1.2.10
langchain-core==1.2.16
langchain-google-genai==4.2.1
google-generativeai==0.8.6
